Rename `filters/builtin.rs` into `filters/core.rs`
