+++ askama/src/filters/mod.rs
@@ -12,7 +12,7 @@
 
 #[cfg(feature = "alloc")]
 mod alloc;
-mod builtin;
+mod core;
 mod escape;
 mod humansize;
 mod indent;
@@ -28,7 +28,7 @@ pub use self::alloc::{
     capitalize, fmt, format, linebreaks, linebreaksbr, lower, lowercase, paragraphbreaks, title,
     titlecase, trim, upper, uppercase, wordcount,
 };
-pub use self::builtin::{PluralizeCount, center, join, pluralize, reject, reject_with, truncate};
+pub use self::core::{PluralizeCount, center, join, pluralize, reject, reject_with, truncate};
 pub use self::escape::{
     AutoEscape, AutoEscaper, Escaper, Html, HtmlSafe, HtmlSafeOutput, MaybeSafe, Safe, Text,
     Unsafe, Writable, WriteWritable, e, escape, safe,
